+++ lib/processor/condition/constructor.go
@@ -55,57 +55,60 @@ var Constructors = map[string]TypeSpec{}
 
 // String constants representing each condition type.
 var (
-	TypeAnd         = "and"
-	TypeBoundsCheck = "bounds_check"
-	TypeCheckField  = "check_field"
-	TypeCount       = "count"
-	TypeJMESPath    = "jmespath"
-	TypeNot         = "not"
-	TypeMetadata    = "metadata"
-	TypeOr          = "or"
-	TypeResource    = "resource"
-	TypeStatic      = "static"
-	TypeText        = "text"
-	TypeXor         = "xor"
+	TypeAnd             = "and"
+	TypeBoundsCheck     = "bounds_check"
+	TypeCheckField      = "check_field"
+	TypeCount           = "count"
+	TypeJMESPath        = "jmespath"
+	TypeNot             = "not"
+	TypeMetadata        = "metadata"
+	TypeOr              = "or"
+	TypeProcessorFailed = "processor_failed"
+	TypeResource        = "resource"
+	TypeStatic          = "static"
+	TypeText            = "text"
+	TypeXor             = "xor"
 )
 
 //------------------------------------------------------------------------------
 
 // Config is the all encompassing configuration struct for all condition types.
 type Config struct {
-	Type        string            `json:"type" yaml:"type"`
-	And         AndConfig         `json:"and" yaml:"and"`
-	BoundsCheck BoundsCheckConfig `json:"bounds_check" yaml:"bounds_check"`
-	CheckField  CheckFieldConfig  `json:"check_field" yaml:"check_field"`
-	Count       CountConfig       `json:"count" yaml:"count"`
-	JMESPath    JMESPathConfig    `json:"jmespath" yaml:"jmespath"`
-	Not         NotConfig         `json:"not" yaml:"not"`
-	Metadata    MetadataConfig    `json:"metadata" yaml:"metadata"`
-	Or          OrConfig          `json:"or" yaml:"or"`
-	Plugin      interface{}       `json:"plugin,omitempty" yaml:"plugin,omitempty"`
-	Resource    string            `json:"resource" yaml:"resource"`
-	Static      bool              `json:"static" yaml:"static"`
-	Text        TextConfig        `json:"text" yaml:"text"`
-	Xor         XorConfig         `json:"xor" yaml:"xor"`
+	Type            string                `json:"type" yaml:"type"`
+	And             AndConfig             `json:"and" yaml:"and"`
+	BoundsCheck     BoundsCheckConfig     `json:"bounds_check" yaml:"bounds_check"`
+	CheckField      CheckFieldConfig      `json:"check_field" yaml:"check_field"`
+	Count           CountConfig           `json:"count" yaml:"count"`
+	JMESPath        JMESPathConfig        `json:"jmespath" yaml:"jmespath"`
+	Not             NotConfig             `json:"not" yaml:"not"`
+	Metadata        MetadataConfig        `json:"metadata" yaml:"metadata"`
+	Or              OrConfig              `json:"or" yaml:"or"`
+	Plugin          interface{}           `json:"plugin,omitempty" yaml:"plugin,omitempty"`
+	ProcessorFailed ProcessorFailedConfig `json:"processor_failed" yaml:"processor_failed"`
+	Resource        string                `json:"resource" yaml:"resource"`
+	Static          bool                  `json:"static" yaml:"static"`
+	Text            TextConfig            `json:"text" yaml:"text"`
+	Xor             XorConfig             `json:"xor" yaml:"xor"`
 }
 
 // NewConfig returns a configuration struct fully populated with default values.
 func NewConfig() Config {
 	return Config{
-		Type:        "text",
-		And:         NewAndConfig(),
-		BoundsCheck: NewBoundsCheckConfig(),
-		CheckField:  NewCheckFieldConfig(),
-		Count:       NewCountConfig(),
-		JMESPath:    NewJMESPathConfig(),
-		Not:         NewNotConfig(),
-		Metadata:    NewMetadataConfig(),
-		Or:          NewOrConfig(),
-		Plugin:      nil,
-		Resource:    "",
-		Static:      true,
-		Text:        NewTextConfig(),
-		Xor:         NewXorConfig(),
+		Type:            "text",
+		And:             NewAndConfig(),
+		BoundsCheck:     NewBoundsCheckConfig(),
+		CheckField:      NewCheckFieldConfig(),
+		Count:           NewCountConfig(),
+		JMESPath:        NewJMESPathConfig(),
+		Not:             NewNotConfig(),
+		Metadata:        NewMetadataConfig(),
+		Or:              NewOrConfig(),
+		Plugin:          nil,
+		ProcessorFailed: NewProcessorFailedConfig(),
+		Resource:        "",
+		Static:          true,
+		Text:            NewTextConfig(),
+		Xor:             NewXorConfig(),
 	}
 }